requirements.txt
langchain
python-dotenv
langchain_text_splitters
langchain-openai
streamlit 
langchain_community
pylance
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette